PDD/requirements.txt
fastapi
uvicorn[standard]
loguru
PyPDF2
scikit-learn
joblib
requests
reportlab
pymongo
firebase-admin
python-multipart
python-dotenv
google-cloud-aiplatform>=1.38.0